\name{rTRMui-package}
\alias{rTRMui-package}
\alias{rTRMui}
\docType{package}
\title{
rTRM shiny UI
}
\description{
A web-based GUI for the rTRM package implemented with the package shiny.
}
\details{
\tabular{ll}{
Package: \tab rTRMui\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-04-07\cr
License: \tab GPL-3\cr
}
}
\author{
Diego Diez

Maintainer: Diego Diez <diego10ruiz@gmail.com>
}

\keyword{ package }